requirements.txt
toml
discordpy
